Refactor: Simplify observability API and remove startup module
